requirements.txt
git+https://github.com/Rapptz/discord.py
async-timeout==3.0.1
Pillow
python-dotenv